+++ src/collider/GeomTree.cpp
@@ -6,13 +6,13 @@
 #include "BVHTree.h"
 #include "Weld.h"
 
-static const unsigned int IGNORE_FLAG = 0x8000;
+static const Uint32 IGNORE_FLAG = 0x8000;
 
 GeomTree::~GeomTree()
 {
 }
 
-GeomTree::GeomTree(const int numVerts, const int numTris, const std::vector<vector3f> &vertices, const Uint16 *indices, const unsigned int *triflags)
+GeomTree::GeomTree(const int numVerts, const int numTris, const std::vector<vector3f> &vertices, const Uint32 *indices, const Uint32 *triflags)
 : m_numVertices(numVerts)
 , m_numTris(numTris)
 , m_vertices(vertices)
@@ -75,7 +75,7 @@ GeomTree::GeomTree(const int numVerts, const int numTris, const std::vector<vect
 	m_radius = 0;
 	for (int i=0; i<numTris; i++)
 	{
-		const unsigned int triflag = m_triFlags[i];
+		const Uint32 triflag = m_triFlags[i];
 		if (triflag < IGNORE_FLAG)
 		{
 			const int vi1 = m_indices[3*i+0];
@@ -103,7 +103,7 @@ GeomTree::GeomTree(const int numVerts, const int numTris, const std::vector<vect
 
 	{
 		Aabb *aabbs = new Aabb[activeTris.size()];
-		for (unsigned int i = 0; i < activeTris.size(); i++)
+		for (Uint32 i = 0; i < activeTris.size(); i++)
 		{
 			const vector3d v1 = vector3d(m_vertices[m_indices[activeTris[i] + 0]]);
 			const vector3d v2 = vector3d(m_vertices[m_indices[activeTris[i] + 1]]);
@@ -190,7 +190,7 @@ GeomTree::GeomTree(Serializer::Reader &rd)
 	const int numIndicies(m_numTris * 3);
 	m_indices.resize(numIndicies);
 	for (Sint32 iIndi = 0; iIndi < numIndicies; ++iIndi) {
-		m_indices[iIndi] = rd.Int16();
+		m_indices[iIndi] = rd.Int32();
 	}
 
 	m_triFlags.resize(m_numTris);
@@ -209,7 +209,7 @@ GeomTree::GeomTree(Serializer::Reader &rd)
 	}
 	// regenerate the aabb data
 	Aabb *aabbs = new Aabb[activeTris.size()];
-	for (unsigned int i = 0; i<activeTris.size(); i++)
+	for (Uint32 i = 0; i<activeTris.size(); i++)
 	{
 		const vector3d v1 = vector3d(m_vertices[m_indices[activeTris[i] + 0]]);
 		const vector3d v2 = vector3d(m_vertices[m_indices[activeTris[i] + 1]]);
@@ -396,7 +396,7 @@ void GeomTree::Save(Serializer::Writer &wr) const
 	}
 
 	for (Sint32 iIndi = 0; iIndi < (m_numTris * 3); ++iIndi) {
-		wr.Int16(m_indices[iIndi]);
+		wr.Int32(m_indices[iIndi]);
 	}
 
 	for (Sint32 iTri = 0; iTri < m_numTris; ++iTri) {
